requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==26.1.0
backoff==2.2.1
blinker==1.9.0
boto3==1.42.89
botocore==1.42.89
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.7
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.9
dagshub-annotation-converter==0.2.0
databricks-sdk==0.102.0
dataclasses-json==0.6.7
dill==0.4.1
dnspython==2.8.0
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.135.3
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.5.0
lxml==6.0.4
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
multidict==6.7.1
mypy_extensions==1.1.0
-e git+https://github.com/rnrahate/Network-Security.git@8ce61827871c23dee6f9a664d038140b4da689b4#egg=network_security
numpy==2.2.6
opentelemetry-api==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==26.1
pandas==2.3.3
pathvalidate==3.3.1
pillow==12.2.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.6
pyaml==26.2.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
pymongo==4.16.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
tenacity==9.1.4
threadpoolctl==3.6.0
tomli==2.4.1
treelib==1.8.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
waitress==3.0.2
wcwidth==0.6.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
